1) Made substantial changes to the GateKeeper and the IExecuteService Interface Changed signature of IExecuteService.ExecuteServiceMethod Old Signature: Object ExecuteServiceMethod( int nTreatyID, string strServiceName, string strServiceMethod Object[] parameters) New Signature: string ExecuteServiceMethod( string strXMLExecRequest )
A request for a service method execution is now specified using XML
see sample file XMLFile1.xml (a formal schema will be defined shortly)
Basically the ExecRequest contains
<TreatyID>,<ServiceName>,<MethodName> and <Parameter>

To Execute a service method the GateKeeper goes thru a 7 stage process
NOTE stages 1 and 2 have not yet been implemented, only stages 3 thru 7

1) Validate schema of strXMLExecRequest - not yet
2) Do Security check
3) Extraction of parameters
4) Resolve Web Service Proxy Location (Proxy may have to be generated at runtime)
5) Runtime Deserializtion of parameters
6) Proxy Method Invocation
7) Serialize and return results

For stage 5 to Deserialize Parameters the GK (after loading the Proxy Assembly and creating the instance)
uses the reflection API to get the metadata about the Method to be invoked (the number of parameters and their types).
The GK will then create an instance of the types and try to Deserialize the
XML passed in with the request into a type instance thereby reconstructing the
array of parameters to be passed to the proxy for method invocation.
This allows the GK to interact with WebSevices that may expect complex types as
input. Once the GK can do the Deserialization (the XML passed in the request must
be correct) then the complex parameter type can be reconstructed and passed to the method.

After method invocation the GK will query the returned object for its type
and attempt to Serialize it to XML and return it as a string. This allows the GK to
interact with WebServices that may return complex types but through the use of reflection
and XML serialization the GK can XML-ify them and return their XML representation.